% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curry_amount_except_pentuple_it.R
\name{curry_amount_except_pentuple_it}
\alias{curry_amount_except_pentuple_it}
\title{Title}
\usage{
curry_amount_except_pentuple_it(initial_amount, customer_name)
}
\arguments{
\item{initial_amount}{A number representing the amount of servings of curry the customer wants. MUST be <= 2.}

\item{customer_name}{A string representing that is the customer's name}
}
\value{
A number that is five times more curry than they asked for. More curry is always good.
}
\description{
Title
}
\examples{
#curry_amount_except_pentuple_it(10, "Zach")
}
